Create structured output script
